more readable file structure + post methode fixes
