--- 1.txt
@@ -1,1 +1,0 @@
-hi
