+++ cgi-bin/test.py
@@ -1,5 +1,16 @@
 #!/usr/bin/env python3
-import os, sys
+import cgi
+import cgitb
+cgitb.enable()
 
-while True:
-    sys.stdin.read();
+print("Content-type: text/html\n")
+print("<html><body>")
+print("<h1>Python CGI Test</h1>")
+
+form = cgi.FieldStorage()
+if "name" in form:
+    print(f"<p>Hello, {form['name'].value}!</p>")
+else:
+    print("<p>No name provided</p>")
+
+print("</body></html>")
