+++ conf/test.conf
@@ -29,5 +29,12 @@ server {
         allowed_methods POST GET;
         upload_dir uploads;
     }
+	
+	location /cgi-bin/ {
+    root ./;
+    allowed_methods GET POST;
+    cgi_path /usr/bin/python3 /usr/bin/php-cgi;
+    cgi_ext .py .php;
+}
 
 }
